+++ TapSDK/TapBootstrapSDK/ITapBootstrap.cs
@@ -6,10 +6,16 @@ namespace TapTap.Bootstrap
     {
         void Init(TapConfig config);
 
+        [Obsolete("已弃用,请调用Login(string[] permissions)")]
         void Login(LoginType loginType, string[] permissions);
 
+        [Obsolete("已弃用,请调用Login(string[] permissions)")]
         void Bind(LoginType loginType, string[] permissions);
 
+        void Login(string[] permissions);
+
+        void GetTestQualification(Action<bool, TapError> action);
+        
         void RegisterUserStatusChangedListener(ITapUserStatusChangedListener listener);
 
         void RegisterLoginResultListener(ITapLoginResultListener listener);