+++ TapSDK/TapBootstrapSDK/LoginType.cs
@@ -2,8 +2,6 @@ namespace TapTap.Bootstrap
 {
     public enum LoginType
     {
-        TAPTAP = 0,
-        APPLE = 1,
-        GUEST = 2
+        TAPTAP = 0
     }
 }