+++ Assets/TapTap/Moment/CHANGELOG.md
@@ -1,5 +1,12 @@
 # ChangeLog
 
+## 2.1.0
+
+### Dependencies
+
+* TapTap.Bootstrap v2.1.0
+* TapTap.Common v2.1.0
+
 ## 2.0.0
 
 ### Dependencies